gui2/unit_attack: Only display active specials on the weapons list
This fixes issues reported on the forums where the weapons list now
includes weapon specials even when they aren't active, a regression from
1.12.x.

However, this introduces a new regression in the form of backstab never
being considered active for the effects of this even when it clearly is
(i.e. the damage stats still reflect it but it isn't listed).