gui/language_selection: Switch languages immediately on item selection
This makes it easier to tell how much of the user interface is
translated to the new language, with the exception of an issue with
Version and Language menu labels in the main menu requiring a manual
refresh.

(This issue could really be fixed by having the title screen dialog
give us a function to call to have it refresh the labels for us. Or
by figuring out some other way to signal it to do so.)