gui: Change syntax used for modifier descriptions in the hotkeys UI
On non-Apple platforms, this makes it so modifier key names are
capitalised. On Apple platforms, this makes the key separator a blank
instead of the + sign, and uses special characters to describe the
keys as is commonly done in software running on Apple platforms.

(The implementation of the method that produces these strings is also
more compact now.)